# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Picpay test Open api",
        "description": "Tiago braga test for picpay.",
        "contact": {
            "email": "tiagofvx@gmail.com"
        },
        "license": {
            "name": "None"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Creates a user.",
                "description": "Creates a user.",
                "operationId": "storeUser",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    },
                                    "cpf": {
                                        "description": "CPF",
                                        "type": "string"
                                    },
                                    "cnpj": {
                                        "description": "CNPJ",
                                        "type": "string"
                                    },
                                    "type": {
                                        "description": "Type",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Invalid data supplied",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login.",
                "description": "Login user.",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Invalid data supplied",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/detail": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Gets the authenticated user",
                "description": "Gets the authenticated user",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transfers": {
            "get": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Get all transfers. Must be authenticated.",
                "description": "Returns list of transfers for the authenticated user.",
                "operationId": "getTransfersIndex",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Stores a transfer. Must be authenticated.",
                "description": "Stores a transfer",
                "operationId": "storeTransfer",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payee_id": {
                                        "description": "The id of the receiver",
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "description": "The amount that the user wants to transfer",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/transfers/{id}": {
            "get": {
                "tags": [
                    "Transfers"
                ],
                "summary": "Get detail about a transfer. Must be authenticated.",
                "description": "Returns a transfer",
                "operationId": "getTransfer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Transfer uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Transfer not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/wallets": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Get list of all Wallets. Must be authenticated.",
                "description": "Returns list of wallets",
                "operationId": "getWalletsIndex",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/wallets/{id}": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Get detail about a wallet. Must be authenticated.",
                "description": "Returns list of wallets",
                "operationId": "getWallet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Wallet id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}